fixed an issue from the updated thread-safing of mechanical feedback that was causing a continue for r-process and age tracers before they were re-coupled

+++ allvars.h
@@ -1961,7 +1961,7 @@ extern struct global_data_all_processes
   int NumFilesWrittenInParallel;	/*!< maximum number of files that may be written simultaneously when
                                      writing/reading restart-files, or when writing snapshot files */
   double BufferSize;		/*!< size of communication buffer in MB */
-  int BunchSize;     	        /*!< number of particles fitting into the buffer in the parallel tree algorithm  */
+  long BunchSize;     	        /*!< number of particles fitting into the buffer in the parallel tree algorithm  */
 
   double PartAllocFactor;	/*!< in order to maintain work-load balance, the particle load will usually
 				   NOT be balanced.  Each processor allocates memory for PartAllocFactor times
